requirements.txt
python-socketio>=5.5.0
eventlet>=0.30.2
pandas>=1.3.0
matplotlib>=3.4.0